embed predict+attention modules+bert
This works ok. The attention is leaky as expected so it doesn't handle long sequences well. with manual attention (on spaces for now, could be spaces and all special chars) it predicts most of the right embeddings but the sequence is easily broken by an early incorrect prediction (learned attention has this same issue). I'll be working on an objective in which the exact sequence of found tokens does not matter as much. maybe cloze LM with a fixed # of masked items.